added focusbot to back end
